Initial commit of Git-bash environment files for MS Windows
This contains the .bashrc and a .bash_aliases file that you can use
to customize your Git-bash environment on MS windows.

The .bashrc allows you to specify an SSH private key that will
be forwarded using ssh-agent.  This can be useful when connecting
to e.g. the Amazon Web Services cloud computing platform.

A README file will be added at a later time.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>